Always send emails in the user's preferred language
Previously, if the email was sent as a result of an action that
didn't require authorization (like password reset), its
language was set to whatever language the initiator uses.

So, if someone who is not me tried to reset my password and their
browser language is English, I'd receive an email in English
even if my preferred language was set to Russian.

This is fixed here.