Bump JS target version to ES2015
It's been around for 10 years already, all major browsers
support it, and those that don't probably also don't
support some web APIs that are currently being used in
Smithereen.

This results in a smaller bundle size and unlocks further
improvements in the bundle size once we modularize the JS code.

Before:
- 175782 bytes for common.js
- 10254 bytes for graffiti.js
- 34550 bytes for graph.js

After:
- 164081 byte for common.js (6.7% smaller)
- 8837 bytes for graffiti.js (13.8% smaller)
- 31971 byte for graph.js (7.5% smaller)