Automatically focus on the confirmation button in ConfirmBox
If we don't do that, the focus remains on the last clicked
link/button. If the user pressed Enter after that,
the 'click' event would be fired again on the same link,
which fucked up the layer hierarchy because we'd try
to push _another_ ConfirmBox to the layer stack.

A nice side effect of this fix is that now it's not
necessary to use the mouse when confirming stuff.
One can just press Enter.

Fixes #149