Delete abandoned files AFTER deleting rows in `media_files`
Make sure that we only delete actual files on disk
if the corresponding rows in the `media_files` table
were successfully deleted.

Otherwise, it is possible that attempting to delete
a row in `media_files` results in a DB constraint
violation, but at this point it is already too late,
since we deleted the actual file from disk.
This results in data loss.

With this fix, if row deletion fails for any reason,
we roll back the state of the `media_files` table
and don't delete any files.